backend/requirements.txt
# FastAPI 和服务器
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# LangChain 核心
langchain==0.3.10
langchain-community==0.3.10
langchain-openai==0.2.14

# Milvus 向量数据库
pymilvus==2.5.1

# Embedding 模型
FlagEmbedding==1.3.3

# 文件解析
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2

# HTTP 请求与网页解析
requests==2.32.3
beautifulsoup4==4.12.3

# 环境变量
python-dotenv==1.0.1
